% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{.apply_color_mapping}
\alias{.apply_color_mapping}
\title{Apply Color Mapping (Categorical or Continuous)}
\usage{
.apply_color_mapping(
  p,
  layoutDf,
  color.by,
  color.palette,
  color.legend.title,
  geom_type = c("color", "fill"),
  unenrichedColor = "#CCCCCC"
)
}
\description{
Apply Color Mapping (Categorical or Continuous)
}
\keyword{internal}
