% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{.apply_theme}
\alias{.apply_theme}
\title{Apply Common Theme to ggraph Plot}
\usage{
.apply_theme(
  p,
  bgColor,
  edgeColor,
  fontFamily,
  nodeSizeRange,
  size.legend.title,
  title,
  subtitle
)
}
\description{
Apply Common Theme to ggraph Plot
}
\keyword{internal}
