% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{.build_base_graph}
\alias{.build_base_graph}
\title{Build Base Graph Structure}
\usage{
.build_base_graph(layoutDf, edgesDf, edgeColor, verbose)
}
\description{
Build Base Graph Structure
}
\keyword{internal}
