% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{.create_label_data}
\alias{.create_label_data}
\title{Create Label Data for Direct Labels}
\usage{
.create_label_data(layoutDf, centroid_threshold = 2)
}
\description{
Create Label Data for Direct Labels
}
\keyword{internal}
