% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{.create_tooltip}
\alias{.create_tooltip}
\title{Create Tooltip HTML for visNetwork}
\usage{
.create_tooltip(
  complexId,
  primaryFunctionalDomain,
  proteinCount,
  proteins,
  color.by = NULL,
  color.value = NULL
)
}
\description{
Create Tooltip HTML for visNetwork
}
\keyword{internal}
