% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{explore}
\alias{explore}
\title{Explore a ComplexMap Object Interactively}
\usage{
explore(cmap, ...)
}
\arguments{
\item{cmap}{A `ComplexMap` object.}

\item{...}{Additional arguments passed to `runComplexMapApp`.}
}
\description{
A convenient alias for `runComplexMapApp()`. Launches the Cytoscape.js
interactive explorer.
}
