% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexmap_S3.R
\name{getEdgeTable}
\alias{getEdgeTable}
\title{Get the Edge Table from a ComplexMap Object}
\usage{
getEdgeTable(cm)
}
\arguments{
\item{cm}{A `ComplexMap` object.}
}
\value{
A `tibble`.
}
\description{
Access the network edge list (weights, similarity types).
}
