% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGmt.R
\name{getExampleGmt}
\alias{getExampleGmt}
\title{Get the Path to an Example GMT File}
\usage{
getExampleGmt()
}
\value{
A character string containing the full path to the example GMT file.
}
\description{
Provides the full system path to an example gene set (GMT)
file included with the ComplexMap package.
}
\details{
The included file is the BioCarta gene set collection from the
Molecular Signatures Database (MSigDB v2025.1). This function makes it easy
to access the file for use in examples and vignettes.
}
\examples{
# Get the path
gmtPath <- getExampleGmt()

# You can then read the file using the path
if (file.exists(gmtPath)) {
  exampleGmt <- getGmtFromFile(gmtPath)
}

}
