% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGmt.R
\name{getGoGmt}
\alias{getGoGmt}
\title{Fetch Gene Ontology (GO) Gene Sets}
\usage{
getGoGmt(speciesDb, ontology = "BP", minGmtSize = 10, maxGmtSize = 500)
}
\arguments{
\item{speciesDb}{An AnnotationDb object for the target species (e.g.,
`org.Hs.eg.db`).}

\item{ontology}{The GO ontology to fetch. One of "BP", "MF", or "CC".}

\item{minGmtSize}{The minimum number of genes for a GO term to be included.}

\item{maxGmtSize}{The maximum number of genes for a GO term to be included.}
}
\value{
A named list where names are GO terms and values are character
  vectors of Entrez IDs.
}
\description{
Fetches GO terms and their associated genes from a Bioconductor
AnnotationDb package.
}
