% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexmap_S3.R
\name{getNodeTable}
\alias{getNodeTable}
\title{Get the Node Table from a ComplexMap Object}
\usage{
getNodeTable(cm)
}
\arguments{
\item{cm}{A `ComplexMap` object.}
}
\value{
A `tibble`.
}
\description{
Access the node attributes (layout, function, specificity scores).
}
