% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{queryMap}
\alias{queryMap}
\title{Query a ComplexMap Object}
\usage{
queryMap(complexMapObject, query, type)
}
\arguments{
\item{complexMapObject}{A `ComplexMap` object.}

\item{query}{Search string.}

\item{type}{"protein", "complex", or "theme".}
}
\value{
A tibble of matching nodes.
}
\description{
Targeted querying of complexes, proteins, or themes.
}
\details{
- `type="protein"`: Regex search for protein members.
- `type="complex"`: Exact match for Complex ID.
- `type="theme"`: Exact match for Theme Label. Requires `summarizeThemes()`
  to be run on the object first with `add_to_object = TRUE`.
}
