% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexmap_S3.R
\name{getEdgeTable}
\alias{getEdgeTable}
\title{Get the Edge Table from a ComplexMap Object}
\usage{
getEdgeTable(cm)
}
\arguments{
\item{cm}{A `ComplexMap` object, typically the output of
`createComplexMap()`.}
}
\value{
A `tibble` containing the edge data.
}
\description{
A simple accessor function to extract the tibble of network edges from a
`ComplexMap` object.
}
\examples{
# Assuming 'myComplexMap' is a valid ComplexMap object
# edgeData <- getEdgeTable(myComplexMap)
}
