% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexmap_S3.R
\name{print.ComplexMap}
\alias{print.ComplexMap}
\title{Print Method for a ComplexMap Object}
\usage{
\method{print}{ComplexMap}(x, ...)
}
\arguments{
\item{x}{A `ComplexMap` object.}

\item{...}{Additional arguments passed to `print` (not used by this method).}
}
\value{
Invisibly returns the original `ComplexMap` object, allowing for
  its use in pipelines.
}
\description{
Provides a user-friendly summary of the `ComplexMap` object when it is
printed to the console. This method is automatically called by the `print`
generic function.
}
